Remove Azure Web App workflow; add Container Apps workflow
This commit removes the workflow configuration for building and deploying a .NET Core application to Azure Web Apps in `blazor-marquitos.yml`.

A new workflow is introduced in `NorteContainerAppsEnvironment.yml` for building and deploying a containerized application to Azure Container Apps. This includes setting up environment variables, defining a job that runs on Ubuntu, and specifying steps for code checkout, .NET SDK setup, Azure login, and container app deployment using a matrix strategy.